groq_translator/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
groq>=0.4.0
pydantic>=1.8.0,<2.0.0
requests>=2.31.0
python-multipart>=0.0.5 